Loaded user id.

+++ src/jvmMain/kotlin/repository/UsersRepo.kt
@@ -10,6 +10,6 @@ import model.User
 interface UsersRepo {
     fun current(): MutableState<User?>
     fun all():List<User>
-    fun addListener(listener: () -> Unit)
+    fun addListener(listener: (Long) -> Unit)
     fun setCurrentUser(user: User)
 }
